Group protractor tests into a single file.
