+++ core/tests/protractor.conf.js
@@ -90,14 +90,11 @@ exports.config = {
     ],
 
     users1: [
-      // 'protractor/userManagement.js',
-      // 'protractor/loginFlow.js',
-      'protractor/subscriptions.js',
-      // 'protractor/preferences.js'
+      'protractor/userManagement.js',
     ],
 
     users2: [
-      'protractor/loginFlow.js',
+      'protractor/subscriptions.js',
     ],
 
     misc: [
